Continuing to refactor into logic and vis. Logic is as finished as it should be at the moment, it updates within the player class like it did before the refactor. Working on getting the collection of entities to draw now, might have to retire the Draw function inside of the entity class and use the Vis one instead, other way around would also work.

+++ androidStudio/app/src/main/java/uk/ac/tees/w9039358/mobileAndGamingDevices/Vector2D.java
@@ -1,19 +1,19 @@
 package uk.ac.tees.w9039358.mobileAndGamingDevices;
 
 public class Vector2D {
-    private int X, Y;
+    private float X, Y;
 
-    Vector2D(int x, int y)
+    Vector2D(float x, float y)
     {
         X = x;
         Y = y;
     }
 
-    public int GetX() {
+    public float GetX() {
         return X;
     }
 
-    public int GetY() {
+    public float GetY() {
         return Y;
     }
